pyproject.toml
[project]
name = "image-metadata-cleaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "pyinstaller>=6.13.0",
    "tkinterdnd2>=0.4.3",
]
